Some fixes
In `6bd0e08fbfac244aab0c62e18c9b9e4b72d2d2f6` I introduced some bugs
:-), this should fix them.

+++ cmk/utils/render.py
@@ -253,7 +253,7 @@ def fmt_number_with_precision(
 
 
 def fmt_bytes(
-    b: int,
+    b: float,
     *,
     unit_prefix_type: type[_ABCUnitPrefixes] = IECUnitPrefixes,
     precision: int = 2,
